blind copy of all libraries

+++ trunk/cprover/src/util/string_container.cpp
@@ -6,9 +6,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 \*******************************************************************/
 
-#include <iostream>
-
-#include <string.h>
+#include <cstring>
 
 #include "string_container.h"
 
@@ -109,7 +107,7 @@ unsigned string_containert::get(const char *s)
   if(it!=hash_table.end())
     return it->second;
 
-  unsigned r=hash_table.size();
+  size_t r=hash_table.size();
 
   // these are stable
   string_list.push_back(std::string(s));
@@ -144,7 +142,7 @@ unsigned string_containert::get(const std::string &s)
   if(it!=hash_table.end())
     return it->second;
 
-  unsigned r=hash_table.size();
+  size_t r=hash_table.size();
 
   // these are stable
   string_list.push_back(s);
